requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pytest
pytest-django